requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
pinecone-client>=3.0.0
sentence-transformers>=2.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
sqlalchemy>=2.0.0
